Restructure viewer files
