+++ docs/viewer/test.ts
@@ -1,7 +1,7 @@
 import { test } from "./test/util/cappucino";
 
-import * as ComposeAlphaTest from "./compose-alpha-test"
-import * as ExampleFileTest from "./example-file-test"
+import * as ComposeAlphaTest from "./test/compose-alpha-test"
+import * as ExampleFileTest from "./test/example-file-test"
 
 ComposeAlphaTest
 ExampleFileTest
