+++ docs/viewer/test/example-file-test.ts
@@ -1,8 +1,8 @@
-import { describe, each, it } from "./test/util/cappucino";
+import { describe, each, it } from "./util/cappucino";
 import { expect } from "chai";
 import * as fs from "fs";
-import { LoadIfcxFile, SchemaValidationError } from "./workflow-alpha";
-import { components } from "../../schema/out/ts/ifcx";
+import { LoadIfcxFile, SchemaValidationError } from "../workflow-alpha";
+import { components } from "../../../schema/out/ts/ifcx";
 
 type IfcxFile = components["schemas"]["IfcxFile"];
